requirements.txt
imglyb==2.1.0
jgo==1.0.5
JPype1==1.4.1
labeling==0.1.13
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.5.0
psutil==5.9.4
pyimagej==1.4.1
python-dateutil==2.8.2
pytz==2023.3
scipy==1.10.1
scyjava==1.9.0
six==1.16.0
tifffile==2023.3.21
xarray==2023.3.0
